requirements.txt
insightface>=0.7.0
opencv-python>=4.7.0.72
numpy>=1.24.3
pillow>=9.5.0

# Dependencias opcionales para desarrollo y construcción
pyinstaller>=5.13.0  # Para crear ejecutables